fix!: agree on integer widths for the presenter overlay setting and CV lock flags
The presenter overlay privacy alert exports took and returned a Swift
Int while Rust passed and read an i32, so on arm64 Swift read the
undefined upper half of the argument register and Rust dropped the
upper half of the result. The raw cm::ffi declarations of apple-cf's
cv_pixel_buffer_{lock,unlock}_base_address passed u32 flags to exports
that take UInt64, with the same effect.

Both presenter overlay exports now use Int32. The setter maps only the
three known raw values and reports anything else instead of ignoring
it; imported NS_ENUMs accept any raw value in init?(rawValue:), so the
check is an explicit switch. The getter passes the raw value through
with a saturating conversion, so Rust rejects an unknown value instead
of reading it as System.

On the Rust side the setter, its builder and the getter return an
SCResult: FeatureNotAvailable where the property doesn't exist (before
macOS 14) instead of dropping the value or reporting System, and
FFIError for an unknown raw value. SCPresenterOverlayAlertSetting gains
from_raw. The lock and unlock declarations take u64 flags.

Tests round-trip every setting through the bridge and through Clone,
check from_raw against unknown values, and drive the raw export with
unknown values to check it refuses them and keeps the previous setting.
None of them needs Screen Recording permission.

Copilot-Session: 61abddcf-4f9d-46ad-ba0d-7a0382752af1

+++ src/cm/ffi.rs
@@ -238,11 +238,11 @@ extern "C" {
     pub fn cv_pixel_buffer_get_bytes_per_row(pixel_buffer: *mut std::ffi::c_void) -> usize;
     pub fn cv_pixel_buffer_lock_base_address(
         pixel_buffer: *mut std::ffi::c_void,
-        flags: u32,
+        flags: u64,
     ) -> i32;
     pub fn cv_pixel_buffer_unlock_base_address(
         pixel_buffer: *mut std::ffi::c_void,
-        flags: u32,
+        flags: u64,
     ) -> i32;
     pub fn cv_pixel_buffer_get_base_address(
         pixel_buffer: *mut std::ffi::c_void,
